代码修改

+++ index-server/src/main/java/com/cmri/um/he/index/market/service/impl/AppMarketAddServiceImpl.java
@@ -35,7 +35,7 @@ public class AppMarketAddServiceImpl implements AppMarketAddService {
             List<Map<String, Object>> list1 = appMarketAddDao.getAppNumber(category, app, month);
             List<Map<String, Object>> list2 = appMarketAddDao.getAppNumber(category, app, defaultTime);
 
-            if ((int) list1.get(0).get("new_user") >= (int) list2.get(0).get("new_user")) {
+            if ((double) list1.get(0).get("new_user") >= (double) list2.get(0).get("new_user")) {
                 list1.get(0).put("new_userSatte", 1);
             } else {
                 list1.get(0).put("new_userSatte", 0);
